% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_function.R
\name{space_function}
\alias{space_function}
\title{A Space Function}
\usage{
space_function(love = TRUE)
}
\arguments{
\item{love}{Do you love space? Defaults to TRUE.}
}
\description{
This function allows you to express your love of space.
}
\examples{
space_function()
}
\keyword{space}
